[skip ci] tRPC shared package
