+++ .changeset/sharp-ducks-peel.md
@@ -1,0 +1,5 @@
+---
+"@saleor/trpc": major
+---
+
+Introduced shared tRPC package. It contains reusable utils that are shared among apps
