+++ .changeset/small-vans-rhyme.md
@@ -1,0 +1,5 @@
+---
+"saleor-app-segment": patch
+---
+
+Extracted some tRPC utilities to shared package
