+++ apps/segment/package.json
@@ -18,6 +18,7 @@
     "@saleor/apps-ui": "workspace:*",
     "@saleor/macaw-ui": "0.8.0-pre.127",
     "@saleor/react-hook-form-macaw": "workspace:*",
+    "@saleor/trpc": "workspace:*",
     "@segment/analytics-node": "^1.1.0",
     "@sentry/nextjs": "7.55.2",
     "@tanstack/react-query": "^4.29.19",
