+++ packages/trpc/.eslintrc
@@ -1,0 +1,4 @@
+{
+  "root": true,
+  "extends": ["saleor"]
+}
