+++ packages/trpc/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@saleor/trpc",
+  "version": "0.0.0",
+  "scripts": {
+    "lint:fix": "eslint --fix ."
+  },
+  "devDependencies": {
+    "@saleor/app-sdk": "0.41.1",
+    "eslint": "8.46.0",
+    "eslint-config-saleor": "workspace:*",
+    "next": "13.4.8",
+    "typescript": "5.1.6"
+  },
+  "main": "index.ts",
+  "peerDependencies": {
+    "@saleor/app-sdk": "0.41.1",
+    "@tanstack/react-query": "^4.29.19",
+    "@trpc/client": "10.34.0",
+    "@trpc/next": "10.34.0",
+    "@trpc/react-query": "10.34.0",
+    "@trpc/server": "10.34.0",
+    "next": "13.3.0",
+    "pino": "^8.14.1",
+    "pino-pretty": "^10.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "urql": "^4.0.4"
+  }
+}
