+++ packages/trpc/turbo.json
@@ -1,0 +1,9 @@
+{
+  "extends": ["//"],
+  "$schema": "https://turbo.build/schema.json",
+  "pipeline": {
+    "build": {
+      "env": ["VERCEL_URL", "PORT"]
+    }
+  }
+}
